refactor(core): extract DeepMap's three cohesive subsystems
DeepMap drops from ~2100 to ~1450 lines by moving three deletion-test-
passing seams into package-private siblings, bodies verbatim:

- TelescopeFixups — the telescope-row overlay subsystem: the two exhaustive
  per-row forward/backward switches, Conditional peeling, ZIP cardinality,
  and overrideTargetField, behind one static wrap(...) entry point.
- ContainerLifts — the container copy/lift Isos (collection/map copy, the
  three lift*IntoTargetRaw shapes) and the java.base allocator tables.
- Placeholders — placeholderIsoFor, defaultAllocatorIso, recursiveDefault,
  primitiveDefault, beanIntermediateAllocatable, and NULLING_ISO.

What remains in DeepMap is the actual engine: populateIso / autoIso /
assembleIso plus the pairing passes. Behavior identical — full :core and
:internal suites green.
